+++ app/src/main/java/hcmute/nhom2/foody/Activity/HistoryActivity.java
@@ -56,7 +56,7 @@ public class HistoryActivity extends AppCompatActivity {
                     cursor.getInt(1),
                     cursor.getInt(2),
                     cursor.getInt(3),
-                    cursor.getInt(4)
+                    cursor.getDouble(4)
             ));
         }
         return list;
